Set ALLOW_UNICODE doctest option.
Just found out about this great optionflag for doctests. It was
brought to `py.test` in version 2.8.0 and can be used like any other
doctest option. I.e. you can enable the option globally in your
`[pytest]` config (where we additionally enable
``NORMALIZE_WHITESPACE`` as there is normally no good reason for not
enabling it) or in a doctest sample like this::

   >>> print(u"blah")  # doctest: +ALLOW_UNICODE
   "blah"

and the test will succeed under Python2.x as well as under Python 3.x.
It generally ignores the difference in ``"sometext"`` and
``u"sometext"``.

This really eases writing doctest samples that work under both,
Python2.x and Python3.x.