Add `get_prefixes()` implementation.
The new function generates nested lists of plain input lists of strings.
The nested lists will put all entries into own lists, that share a
common prefix (which will be first element).

In fact these nested lists can be interpret as binary trees with
prefixed terms following on the left of a node while not prefixed terms
follow on the right:

  get_prefixes(["a", "aa", "aaa", "ab", "b"])
  -> [["a", ["aa", ["aaa"]], "ab"], ["b"]]

which reads as tree:

                        a
                       / \
                      aa  b
                      /\
                    aaa ab

which means: "a" is a prefix of "aa", "aaa", "ab" (all nodes on its
left) but not a prefix of "b".