Set env var `LANG` in tox.ini.
This seems to be the only reliable way to set a certain, foreseeable env
var when testing.

The `LANG` var influences the automatically picked encoding of opened
files. With some encoding set to something like `xx_XX.utf-8`, for
instance, files are opened using 'utf-8'.

With `LANG` unset, we therefore get failing tests that do not fail with
`LANG` set.